+++ source/extensions/omni.isaac.ros2_bridge/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [2.3.6] - 2023-10-09
 ### Fixed
 - Joint State Subscriber supports input state messages with only one mode of drive specified
+- camera noise example errors
 
 ## [2.3.5] - 2023-10-09
 ### Fixed